Crate matrix_sdk_indexeddb

Structs

IndexeddbCryptoStore

An in-memory only store that will forget all the E2EE key once it’s dropped.

IndexeddbStateStore

IndexeddbStateStoreBuilder

Builder for IndexeddbStateStoreBuilderConfig.

Enums

IndexeddbCryptoStoreError

IndexeddbStateStoreError

MigrationConflictStrategy

Sometimes Migrations can’t proceed without having to drop existing data. This allows you to configure, how these cases should be handled.

OpenStoreError

All the errors that can occur when opening an IndexedDB store.

Functions

make_store_config

Create a [StoreConfig] with an opened indexeddb IndexeddbStateStore that uses the given name and passphrase. If encryption is enabled, a IndexeddbCryptoStore with the same parameters is also opened.